refactor: remove integrations.models shim, import config_models directly
