+++ integrations/honeycomb/client.py
@@ -8,7 +8,7 @@ from typing import Any
 
 import httpx
 
-from integrations.models import HoneycombIntegrationConfig
+from integrations.config_models import HoneycombIntegrationConfig
 from integrations.probes import ProbeResult
 from platform.observability.service_errors import capture_service_error
