+++ integrations/jira/client.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import httpx
 
-from integrations.models import JiraIntegrationConfig
+from integrations.config_models import JiraIntegrationConfig
 from platform.observability.service_errors import capture_service_error
 
 logger = logging.getLogger(__name__)
