--- integrations/models.py
@@ -1,87 +1,0 @@
-"""Backward-compatible re-exports for integration models.
-
-Import directly from config_models or effective_models in new code.
-This shim exists only so existing callers don't need a mass update.
-"""
-
-from integrations.config_models import (
-    AlertmanagerIntegrationConfig,
-    ArgoCDIntegrationConfig,
-    AWSIntegrationConfig,
-    AWSStaticCredentials,
-    AzureSQLIntegrationConfig,
-    BetterStackIntegrationConfig,
-    CoralogixIntegrationConfig,
-    DatadogIntegrationConfig,
-    DiscordBotConfig,
-    GitLabIntegrationConfig,
-    GoogleDocsIntegrationConfig,
-    GrafanaIntegrationConfig,
-    GroundcoverIntegrationConfig,
-    HelmIntegrationConfig,
-    HoneycombIntegrationConfig,
-    IncidentIoIntegrationConfig,
-    JiraIntegrationConfig,
-    MariaDBIntegrationConfig,
-    MongoDBAtlasIntegrationConfig,
-    MongoDBIntegrationConfig,
-    MySQLIntegrationConfig,
-    NotionIntegrationConfig,
-    OpsGenieIntegrationConfig,
-    PostgreSQLIntegrationConfig,
-    PrefectIntegrationConfig,
-    RabbitMQIntegrationConfig,
-    RedisIntegrationConfig,
-    SlackBotConfig,
-    SlackWebhookConfig,
-    SplunkIntegrationConfig,
-    TelegramBotConfig,
-    TracerIntegrationConfig,
-    VercelIntegrationConfig,
-    VictoriaLogsIntegrationConfig,
-)
-from integrations.effective_models import (
-    EffectiveIntegrationEntry,
-    EffectiveIntegrations,
-    IntegrationInstance,
-)
-
-__all__ = [
-    "AWSIntegrationConfig",
-    "AWSStaticCredentials",
-    "AlertmanagerIntegrationConfig",
-    "ArgoCDIntegrationConfig",
-    "AzureSQLIntegrationConfig",
-    "BetterStackIntegrationConfig",
-    "CoralogixIntegrationConfig",
-    "DatadogIntegrationConfig",
-    "DiscordBotConfig",
-    "EffectiveIntegrationEntry",
-    "EffectiveIntegrations",
-    "GitLabIntegrationConfig",
-    "GoogleDocsIntegrationConfig",
-    "GrafanaIntegrationConfig",
-    "GroundcoverIntegrationConfig",
-    "HelmIntegrationConfig",
-    "HoneycombIntegrationConfig",
-    "IncidentIoIntegrationConfig",
-    "IntegrationInstance",
-    "JiraIntegrationConfig",
-    "MariaDBIntegrationConfig",
-    "MongoDBAtlasIntegrationConfig",
-    "MongoDBIntegrationConfig",
-    "MySQLIntegrationConfig",
-    "NotionIntegrationConfig",
-    "OpsGenieIntegrationConfig",
-    "PostgreSQLIntegrationConfig",
-    "PrefectIntegrationConfig",
-    "RabbitMQIntegrationConfig",
-    "RedisIntegrationConfig",
-    "SlackBotConfig",
-    "SlackWebhookConfig",
-    "SplunkIntegrationConfig",
-    "TelegramBotConfig",
-    "TracerIntegrationConfig",
-    "VercelIntegrationConfig",
-    "VictoriaLogsIntegrationConfig",
-]
